fix dev script
