+++ src/entries/popup/pages/_playgrounds/default.tsx
@@ -1,12 +1,11 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React from 'react';
 import { chain, useAccount, useBalance } from 'wagmi';
 import { useFirstTransactionTimestamp } from '~/core/resources/transactions';
 import { Storage } from '~/core/storage';
 import { Box, Text } from '~/design-system';
+import { InjectToggle } from '../../components/InjectToggle';
 
 export function Default() {
-  const [status, setStatus] = useState(0);
-
   const { address } = useAccount();
   const { data: mainnetBalance } = useBalance({
     addressOrName: address,
@@ -19,21 +18,6 @@ export function Default() {
   const { data: firstTransactionTimestamp } = useFirstTransactionTimestamp({
     address,
   });
-
-  const switchInjection = useCallback(async () => {
-    const shouldInject = (await Storage.get('inject')) === true;
-    const newVal = !shouldInject;
-    Storage.set('inject', newVal);
-    setStatus(newVal ? 1 : 0);
-  }, []);
-
-  useEffect(() => {
-    const init = async () => {
-      const shouldInject = (await Storage.get('inject')) === true;
-      setStatus(shouldInject ? 1 : 0);
-    };
-    init();
-  }, []);
 
   return (
     <Box display="flex" flexDirection="column" gap="24px" padding="20px">
@@ -54,31 +38,7 @@ export function Default() {
           </Text>
         )}
       </Box>
-      <Box display="flex" flexDirection="row" gap="8px">
-        <Text size="17pt" weight="bold" color="labelTertiary">
-          Injecting?
-        </Text>
-        <Text
-          size="17pt"
-          weight="bold"
-          color={status ? 'green' : 'red'}
-          testId="injection-status"
-        >
-          {status ? 'YES' : 'NO'}
-        </Text>
-      </Box>
-      <Box
-        as="button"
-        id="injection-button"
-        background="surfaceSecondary"
-        onClick={switchInjection}
-        padding="16px"
-        style={{ borderRadius: 999 }}
-      >
-        <Text color="labelSecondary" size="15pt" weight="bold">
-          TURN {status ? 'OFF' : 'ON'}
-        </Text>
-      </Box>
+      <InjectToggle />
       <Box
         as="button"
         background="surfaceSecondary"
