+++ src/entries/popup/pages/index.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
 import { Box, Text } from '~/design-system';
+import { InjectToggle } from '../components/InjectToggle';
 
 export function Index() {
   return (
-    <Box padding="20px">
-      <Text color="label" size="44pt" weight="heavy">
-        rainbo
+    <Box display="flex" flexDirection="column" gap="24px" padding="20px">
+      <Text as="h1" size="20pt" weight="bold">
+        Rainbow
       </Text>
+      <InjectToggle />
     </Box>
   );
 }
